--- data_model/local_backed/constants.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright (c) Facebook, Inc. and its affiliates.
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-import os.path
-
-# Contains important data model constants
-# TODO update the save dir location, allow setting it globally?
-SAVE_DIR_LOCATION = os.path.expanduser("~/mephisto/data/")
